Extract one common mocking class from unit tests
### What changes are proposed in this pull request?

1. The `CountingNoopFileSystemMaster` is extracted from `UfsJournalTest`
and `RaftJournalTest` into a single class
2. Some RAFT tests are moved from core-common to core-master to use the
extracted class

### Why are the changes needed?

Resolves
https://github.com/Alluxio/alluxio/pull/16060#discussion_r951581855

### Does this PR introduce any user facing changes?

NA

pr-link: Alluxio/alluxio#16108
