+++ core/server/master/src/test/java/alluxio/master/journal/raft/RaftJournalTest.java
@@ -16,6 +16,7 @@ import alluxio.conf.PropertyKey;
 import alluxio.grpc.QuorumServerInfo;
 import alluxio.master.NoopMaster;
 import alluxio.master.journal.CatchupFuture;
+import alluxio.master.journal.CountingNoopFileSystemMaster;
 import alluxio.master.journal.JournalContext;
 import alluxio.proto.journal.File;
 import alluxio.proto.journal.Journal;
@@ -40,7 +41,6 @@ import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
-import java.util.NoSuchElementException;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ForkJoinPool;
 import java.util.stream.Collectors;
@@ -88,7 +88,7 @@ public class RaftJournalTest {
   @Test
   public void writeJournal() throws Exception {
     // Create a counting master implementation that counts how many journal entries it processed.
-    CountingDummyFileSystemMaster countingMaster = new CountingDummyFileSystemMaster();
+    CountingNoopFileSystemMaster countingMaster = new CountingNoopFileSystemMaster();
     mFollowerJournalSystem.createJournal(countingMaster);
 
     // Create entries on the leader journal context.
@@ -124,7 +124,7 @@ public class RaftJournalTest {
 
     RaftJournalSystem newJs = createNewJournalSystem(mLeaderJournalSystem);
     // Create a counting master implementation that counts how many journal entries it processed.
-    CountingDummyFileSystemMaster countingMaster = new CountingDummyFileSystemMaster();
+    CountingNoopFileSystemMaster countingMaster = new CountingNoopFileSystemMaster();
     newJs.createJournal(countingMaster);
     newJs.start();
 
@@ -144,7 +144,7 @@ public class RaftJournalTest {
   @Test
   public void suspendCatchupResume() throws Exception {
     // Create a counting master implementation that counts how many journal entries it processed.
-    CountingDummyFileSystemMaster countingMaster = new CountingDummyFileSystemMaster();
+    CountingNoopFileSystemMaster countingMaster = new CountingNoopFileSystemMaster();
     mFollowerJournalSystem.createJournal(countingMaster);
 
     // Suspend follower journal system.
@@ -202,7 +202,7 @@ public class RaftJournalTest {
   @Test
   public void suspendSnapshotRestart() throws Exception {
     // Create a counting master implementation that counts how many journal entries it processed.
-    CountingDummyFileSystemMaster countingMaster = new CountingDummyFileSystemMaster();
+    CountingNoopFileSystemMaster countingMaster = new CountingNoopFileSystemMaster();
     mFollowerJournalSystem.createJournal(countingMaster);
 
     final int entryCount = 10;
@@ -248,7 +248,7 @@ public class RaftJournalTest {
   @Test
   public void catchUpInSteps() throws Exception {
     // Create a counting master implementation that counts how many journal entries it processed.
-    CountingDummyFileSystemMaster countingMaster = new CountingDummyFileSystemMaster();
+    CountingNoopFileSystemMaster countingMaster = new CountingNoopFileSystemMaster();
     mFollowerJournalSystem.createJournal(countingMaster);
 
     // Suspend follower journal system.
@@ -292,7 +292,7 @@ public class RaftJournalTest {
   @Test
   public void subsequentCatchups() throws Exception {
     // Create a counting master implementation that counts how many journal entries it processed.
-    CountingDummyFileSystemMaster countingMaster = new CountingDummyFileSystemMaster();
+    CountingNoopFileSystemMaster countingMaster = new CountingNoopFileSystemMaster();
     mFollowerJournalSystem.createJournal(countingMaster);
 
     // Suspend follower journal system.
@@ -325,7 +325,7 @@ public class RaftJournalTest {
   public void gainPrimacyAfterSuspend() throws Exception {
 
     // Create a counting master implementation that counts how many journal entries it processed.
-    CountingDummyFileSystemMaster countingMaster = new CountingDummyFileSystemMaster();
+    CountingNoopFileSystemMaster countingMaster = new CountingNoopFileSystemMaster();
     mFollowerJournalSystem.createJournal(countingMaster);
 
     // Suspend follower journal system.
@@ -357,7 +357,7 @@ public class RaftJournalTest {
   @Test
   public void gainPrimacyAfterCatchup() throws Exception {
     // Create a counting master implementation that counts how many journal entries it processed.
-    CountingDummyFileSystemMaster countingMaster = new CountingDummyFileSystemMaster();
+    CountingNoopFileSystemMaster countingMaster = new CountingNoopFileSystemMaster();
     mFollowerJournalSystem.createJournal(countingMaster);
 
     // Suspend follower journal system.
@@ -409,7 +409,7 @@ public class RaftJournalTest {
   @Test
   public void gainPrimacyDuringCatchup() throws Exception {
     // Create a counting master implementation that counts how many journal entries it processed.
-    CountingDummyFileSystemMaster countingMaster = new CountingDummyFileSystemMaster();
+    CountingNoopFileSystemMaster countingMaster = new CountingNoopFileSystemMaster();
     mFollowerJournalSystem.createJournal(countingMaster);
 
     // Using a large entry count for catching transition while in-progress.
@@ -458,7 +458,7 @@ public class RaftJournalTest {
   @Test
   public void catchupCorruptedEntry() throws Exception {
     // Create a counting master implementation that counts how many journal entries it processed.
-    CountingDummyFileSystemMaster countingMaster = new CountingDummyFileSystemMaster();
+    CountingNoopFileSystemMaster countingMaster = new CountingNoopFileSystemMaster();
     mFollowerJournalSystem.createJournal(countingMaster);
     final int entryCount = 3;
     // Suspend follower journal system.
@@ -500,7 +500,7 @@ public class RaftJournalTest {
     });
 
     Assert.assertTrue(exception.getMessage()
-        .contains(CountingDummyFileSystemMaster.ENTRY_DOES_NOT_EXIST));
+        .contains(CountingNoopFileSystemMaster.ENTRY_DOES_NOT_EXIST));
   }
 
   /**
@@ -605,64 +605,5 @@ public class RaftJournalTest {
       socket.close();
     }
     return ports;
-  }
-
-  /**
-   * Used to validate journal apply counts to master.
-   */
-  static class CountingDummyFileSystemMaster extends NoopMaster {
-    public static final String ENTRY_DOES_NOT_EXIST = "The entry to delete does not exist!";
-
-    /** Tracks how many entries have been applied to master. */
-    private long mApplyCount = 0;
-    /** Artificial delay to emulate processing time while applying entries. */
-    private long mApplyDelay = -1;
-
-    @Override
-    public boolean processJournalEntry(Journal.JournalEntry entry) {
-      if (mApplyDelay != -1) {
-        try {
-          Thread.sleep(mApplyDelay);
-        } catch (Exception e) {
-          // do nothing.
-        }
-      }
-      mApplyCount++;
-      // Throw error on a special entry
-      if (entry.hasDeleteFile()) {
-        throw new NoSuchElementException(ENTRY_DOES_NOT_EXIST);
-      }
-      return true;
-    }
-
-    /**
-     * Sets an artificial delay for each apply call.
-     *
-     * @param timeMs delay in ms
-     */
-    public void setApplyDelay(long timeMs) {
-      mApplyDelay = timeMs;
-    }
-
-    @Override
-    public void resetState() {
-      mApplyCount = 0;
-    }
-
-    /**
-     * @return how many entries are applied
-     */
-    public long getApplyCount() {
-      return mApplyCount;
-    }
-
-    @Override
-    public String getName() {
-      /*
-        RaftJournalWriter doesn't accept empty journal entries. FileSystemMaster is returned here
-        according to injected entry type during the test.
-       */
-      return "FileSystemMaster";
-    }
   }
 }
