+++ core/server/master/src/test/java/alluxio/master/journal/ufs/UfsJournalTest.java
@@ -15,8 +15,8 @@ import static org.junit.Assert.assertThrows;
 import static org.junit.Assert.assertTrue;
 
 import alluxio.exception.status.UnavailableException;
-import alluxio.master.NoopMaster;
 import alluxio.master.journal.CatchupFuture;
+import alluxio.master.journal.CountingNoopFileSystemMaster;
 import alluxio.proto.journal.Journal;
 import alluxio.util.CommonUtils;
 import alluxio.util.URIUtils;
@@ -32,7 +32,6 @@ import org.junit.rules.TemporaryFolder;
 import java.io.File;
 import java.net.URI;
 import java.util.Collections;
-import java.util.NoSuchElementException;
 
 /**
  * Unit tests for {@link UfsJournal}.
@@ -45,14 +44,14 @@ public final class UfsJournalTest {
   public ExpectedException mThrown = ExpectedException.none();
 
   private UfsJournal mJournal;
-  private CountingNoopMaster mCountingNoopMaster;
+  private CountingNoopFileSystemMaster mFileSystemMaster;
 
   @Before
   public void before() throws Exception {
-    mCountingNoopMaster = new CountingNoopMaster();
+    mFileSystemMaster = new CountingNoopFileSystemMaster();
     mJournal =
         new UfsJournal(URIUtils.appendPathOrDie(new URI(mFolder.newFolder().getAbsolutePath()),
-            "FileSystemMaster"), mCountingNoopMaster, 0, Collections::emptySet);
+            "FileSystemMaster"), mFileSystemMaster, 0, Collections::emptySet);
   }
 
   /**
@@ -111,7 +110,7 @@ public final class UfsJournalTest {
     mJournal.gainPrimacy();
 
     // Create a counting master implementation that counts how many journal entries it processed.
-    CountingNoopMaster countingMaster = new CountingNoopMaster();
+    CountingNoopFileSystemMaster countingMaster = new CountingNoopFileSystemMaster();
     // Find journal base path for standby journal to consume the same journal files.
     String parentPath = new File(mJournal.getLocation().getPath()).getParent();
     UfsJournal standbyJournal =
@@ -166,12 +165,12 @@ public final class UfsJournalTest {
     // Create a new Journal with the same journal path
     String parentPath = new File(mJournal.getLocation().getPath()).getParent();
     mJournal =
-        new UfsJournal(new URI(parentPath), mCountingNoopMaster, 0, Collections::emptySet);
+        new UfsJournal(new URI(parentPath), mFileSystemMaster, 0, Collections::emptySet);
     mJournal.start();
     CommonUtils.waitFor("catchup done", () -> mJournal.getCatchupState()
             == UfsJournalCheckpointThread.CatchupState.DONE,
         WaitForOptions.defaults().setTimeoutMs(6000));
-    Assert.assertEquals(entryCount, mCountingNoopMaster.getApplyCount());
+    Assert.assertEquals(entryCount, mFileSystemMaster.getApplyCount());
   }
 
   @Test
@@ -185,14 +184,14 @@ public final class UfsJournalTest {
     }
     writer.close();
     mJournal.signalLosePrimacy();
-    Assert.assertEquals(0, mCountingNoopMaster.getApplyCount());
+    Assert.assertEquals(0, mFileSystemMaster.getApplyCount());
     mJournal.awaitLosePrimacy();
     // When master steps down, it should start catching up
     CommonUtils.waitFor("catchup done", () -> mJournal.getCatchupState()
             == UfsJournalCheckpointThread.CatchupState.DONE,
         WaitForOptions.defaults().setTimeoutMs(6000));
     // check if logs are applied
-    Assert.assertEquals(entryCount, mCountingNoopMaster.getApplyCount());
+    Assert.assertEquals(entryCount, mFileSystemMaster.getApplyCount());
   }
 
   @Test
@@ -201,7 +200,7 @@ public final class UfsJournalTest {
     mJournal.gainPrimacy();
 
     // Create a counting master implementation that counts how many journal entries it processed.
-    CountingNoopMaster countingMaster = new CountingNoopMaster();
+    CountingNoopFileSystemMaster countingMaster = new CountingNoopFileSystemMaster();
     // Find journal base path for standby journal to consume the same journal files.
     String parentPath = new File(mJournal.getLocation().getPath()).getParent();
     UfsJournal standbyJournal =
@@ -236,7 +235,7 @@ public final class UfsJournalTest {
     mJournal.gainPrimacy();
 
     // Create a counting master implementation that counts how many journal entries it processed.
-    CountingNoopMaster countingMaster = new CountingNoopMaster();
+    CountingNoopFileSystemMaster countingMaster = new CountingNoopFileSystemMaster();
     // Find journal base path for standby journal to consume the same journal files.
     String parentPath = new File(mJournal.getLocation().getPath()).getParent();
     UfsJournal standbyJournal =
@@ -269,7 +268,7 @@ public final class UfsJournalTest {
     mJournal.gainPrimacy();
 
     // Create a counting master implementation that counts how many journal entries it processed.
-    CountingNoopMaster countingMaster = new SleepingCountingMaster(50);
+    CountingNoopFileSystemMaster countingMaster = CountingNoopFileSystemMaster.withApplyDelay(50);
     // Find journal base path for standby journal to consume the same journal files.
     String parentPath = new File(mJournal.getLocation().getPath()).getParent();
     UfsJournal standbyJournal =
@@ -304,7 +303,7 @@ public final class UfsJournalTest {
     mJournal.gainPrimacy();
 
     // Create a counting master implementation that counts how many journal entries it processed.
-    CountingNoopMaster countingMaster = new CountingNoopMaster();
+    CountingNoopFileSystemMaster countingMaster = new CountingNoopFileSystemMaster();
     // Find journal base path for standby journal to consume the same journal files.
     String parentPath = new File(mJournal.getLocation().getPath()).getParent();
     UfsJournal standbyJournal =
@@ -340,7 +339,7 @@ public final class UfsJournalTest {
     mJournal.gainPrimacy();
 
     // Create a counting master implementation that counts how many journal entries it processed.
-    CountingNoopMaster countingMaster = new CountingNoopMaster();
+    CountingNoopFileSystemMaster countingMaster = new CountingNoopFileSystemMaster();
     // Find journal base path for standby journal to consume the same journal files.
     String parentPath = new File(mJournal.getLocation().getPath()).getParent();
     UfsJournal standbyJournal =
@@ -350,8 +349,6 @@ public final class UfsJournalTest {
     // Suspend standby journal.
     standbyJournal.suspend();
 
-    // Write many entries to guarantee that advancing will be in progress
-    // when gainPrimacy() is called.
     int entryCount = 10;
     for (int i = 0; i < entryCount; i++) {
       mJournal.write(Journal.JournalEntry.getDefaultInstance());
@@ -367,62 +364,13 @@ public final class UfsJournalTest {
     mJournal.write(corruptedEntry);
     mJournal.flush();
 
-    // Set delay for each internal processing of entries before initiating catch-up.
+    // The standby catchup thread will crash and the exception will be
+    // recovered on waitTermination()
     RuntimeException exception = assertThrows(RuntimeException.class, () -> {
       CatchupFuture future = standbyJournal.catchup(entryCount);
       future.waitTermination();
     });
     assertTrue(exception.getMessage()
-        .contains(CountingNoopMaster.ENTRY_DOES_NOT_EXIST));
-  }
-
-  /**
-   * Used to validate journal apply counts to master.
-   */
-  class CountingNoopMaster extends NoopMaster {
-    public static final String ENTRY_DOES_NOT_EXIST = "The entry to delete does not exist!";
-
-    /** Tracks how many entries have been applied to master. */
-    private long mApplyCount = 0;
-
-    @Override
-    public boolean processJournalEntry(Journal.JournalEntry entry) {
-      mApplyCount++;
-      // Throw error on a special entry
-      if (entry.hasDeleteFile()) {
-        throw new NoSuchElementException(ENTRY_DOES_NOT_EXIST);
-      }
-      return true;
-    }
-
-    /**
-     * @return how many entries are applied
-     */
-    public long getApplyCount() {
-      return mApplyCount;
-    }
-  }
-
-  /**
-   * A {@link CountingNoopMaster} with simulated delay per each journal apply.
-   */
-  class SleepingCountingMaster extends CountingNoopMaster {
-    /** How much to sleep per each apply. */
-    private long mSleepMs;
-
-    public SleepingCountingMaster(long sleepMs) {
-      mSleepMs = sleepMs;
-    }
-
-    @Override
-    public boolean processJournalEntry(Journal.JournalEntry entry) {
-      try {
-        Thread.sleep(mSleepMs);
-      } catch (Exception e) {
-        // Do not interfere with interrupt handling.
-        Thread.currentThread().interrupt();
-      }
-      return super.processJournalEntry(entry);
-    }
+        .contains(CountingNoopFileSystemMaster.ENTRY_DOES_NOT_EXIST));
   }
 }
